Langchain/requirements.txt
# LangChain Core Dependencies (updated for Python 3.13)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-ollama>=0.1.0  # For local Ollama models

# OpenAI (for OpenAI models)
openai>=1.30.0

# Google Generative AI (for Gemini models)
google-generativeai>=0.5.0

# Vector Store Dependencies
chromadb>=0.5.0
faiss-cpu>=1.9.0

# Document Loaders
pypdf>=4.0.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.7.0

# Jupyter for interactive learning (optional)
jupyter>=1.0.0
ipykernel>=6.29.0
